+++ dev/mcp-server-examples/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @transcend-io/mcp-server-examples
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies [7e7d797]
+- Updated dependencies [7e7d797]
+  - @transcend-io/mcp-server-base@1.9.0
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies [bccab7e]
+  - @transcend-io/mcp-server-base@1.8.1
+
 ## 0.0.17
 
 ### Patch Changes